update -> create api: clear items for cart, remove list item for cart & optimize productService

+++ src/main/java/com/hdshop/dto/cart/CartResponse.java
@@ -17,6 +17,6 @@ public class CartResponse {
     private Long id;
     private Boolean isDeleted;
     private List<CartItemResponse> cartItems;
-    private Integer totalItems;
+    private long totalItems;
     private BigDecimal totalPrice;
 }
